fix requiredSalary type

+++ src/main/java/methods/InstantCreditCheck.java
@@ -3,7 +3,7 @@ package methods;
 import java.util.Scanner;
 
 public class InstantCreditCheck {
-    static int requiredSalary = 25000;
+    static double requiredSalary = 25000;
     static int requiredCreditScore = 700;
     static Scanner scanner = new Scanner(System.in);
